Revert "feat(step-generation): raise timeline error when pipetting E/W of H-S with multi channel pipette"
This reverts commit ab7aaf9ea4a986be2703e686aba153a77b7fc633.
