+++ protocol-designer/src/localization/en/alert.json
@@ -131,10 +131,6 @@
       "HEATER_SHAKER_NORTH_SOUTH_EAST_WEST_SHAKING": {
         "title": "The Heater-Shaker is shaking",
         "body": "Pipettes cannot access labware on or adjacent to the Heater-Shaker while it is shaking. Create a step before this one that deactivates the shaker."
-      },
-      "HEATER_SHAKER_EAST_WEST_MULTI_CHANNEL": {
-        "title": "8-Channel pipette cannot access labware",
-        "body": "8-Channel pipettes cannot access labware or tip racks to the left or right of a Heater-Shaker GEN1 module. Move labware to a different slot to access it with an 8-Channel pipette."
       }
     },
     "warning": {
