+++ step-generation/src/__tests__/dispense.test.ts
@@ -5,7 +5,6 @@ import {
   pipetteIntoHeaterShakerWhileShaking,
   getIsHeaterShakerEastWestWithLatchOpen,
   pipetteAdjacentHeaterShakerWhileShaking,
-  getIsHeaterShakerEastWestMultiChannelPipette,
 } from '../utils'
 import {
   getInitialRobotStateStandard,
@@ -21,7 +20,10 @@ import { InvariantContext, RobotState } from '../types'
 import type { AspDispAirgapParams as V3AspDispAirgapParams } from '@opentrons/shared-data/protocol/types/schemaV3'
 
 jest.mock('../utils/thermocyclerPipetteCollision')
-jest.mock('../utils/heaterShakerCollision')
+jest.mock('../utils/pipetteIntoHeaterShakerLatchOpen')
+jest.mock('../utils/pipetteIntoHeaterShakerWhileShaking')
+jest.mock('../utils/getIsHeaterShakerEastWestWithLatchOpen')
+jest.mock('../utils/pipetteAdjacentHeaterShakerWhileShaking')
 
 const mockThermocyclerPipetteCollision = thermocyclerPipetteCollision as jest.MockedFunction<
   typeof thermocyclerPipetteCollision
@@ -34,9 +36,6 @@ const mockPipetteIntoHeaterShakerWhileShaking = pipetteIntoHeaterShakerWhileShak
 >
 const mockGetIsHeaterShakerEastWestWithLatchOpen = getIsHeaterShakerEastWestWithLatchOpen as jest.MockedFunction<
   typeof getIsHeaterShakerEastWestWithLatchOpen
->
-const mockGetIsHeaterShakerEastWestMultiChannelPipette = getIsHeaterShakerEastWestMultiChannelPipette as jest.MockedFunction<
-  typeof getIsHeaterShakerEastWestMultiChannelPipette
 >
 const mockPipetteAdjacentHeaterShakerWhileShaking = pipetteAdjacentHeaterShakerWhileShaking as jest.MockedFunction<
   typeof pipetteAdjacentHeaterShakerWhileShaking
@@ -171,7 +170,7 @@ describe('dispense', () => {
       when(mockGetIsHeaterShakerEastWestWithLatchOpen)
         .calledWith(
           robotStateWithTip.modules,
-          robotStateWithTip.labware[SOURCE_LABWARE].slot
+          robotStateWithTip.labware[SOURCE_LABWARE]
         )
         .mockReturnValue(true)
 
@@ -181,26 +180,11 @@ describe('dispense', () => {
         type: 'HEATER_SHAKER_EAST_WEST_LATCH_OPEN',
       })
     })
-    it('should return an error when dispensing east/west of a heater shaker with a multi channel pipette', () => {
-      when(mockGetIsHeaterShakerEastWestMultiChannelPipette)
-        .calledWith(
-          robotStateWithTip.modules,
-          robotStateWithTip.labware[SOURCE_LABWARE].slot,
-          expect.anything()
-        )
-        .mockReturnValue(true)
-
-      const result = dispense(params, invariantContext, robotStateWithTip)
-      expect(getErrorResult(result).errors).toHaveLength(1)
-      expect(getErrorResult(result).errors[0]).toMatchObject({
-        type: 'HEATER_SHAKER_EAST_WEST_MULTI_CHANNEL',
-      })
-    })
     it('should return an error when dispensing north/south/east/west of a heater shaker while it is shaking', () => {
       when(mockPipetteAdjacentHeaterShakerWhileShaking)
         .calledWith(
           robotStateWithTip.modules,
-          robotStateWithTip.labware[SOURCE_LABWARE].slot
+          robotStateWithTip.labware[SOURCE_LABWARE]
         )
         .mockReturnValue(true)
