+++ step-generation/src/commandCreators/atomic/aspirate.ts
@@ -7,7 +7,6 @@ import {
   pipetteIntoHeaterShakerWhileShaking,
   getIsHeaterShakerEastWestWithLatchOpen,
   pipetteAdjacentHeaterShakerWhileShaking,
-  getIsHeaterShakerEastWestMultiChannelPipette,
 } from '../../utils'
 import type { CreateCommand } from '@opentrons/shared-data'
 import type { AspirateParams } from '@opentrons/shared-data/protocol/types/schemaV3'
@@ -96,7 +95,7 @@ export const aspirate: CommandCreator<AspirateParams> = (
   if (
     pipetteAdjacentHeaterShakerWhileShaking(
       prevRobotState.modules,
-      prevRobotState.labware[labware]?.slot
+      prevRobotState.labware[labware]
     )
   ) {
     errors.push(errorCreators.heaterShakerNorthSouthEastWestShaking())
@@ -104,21 +103,12 @@ export const aspirate: CommandCreator<AspirateParams> = (
   if (
     getIsHeaterShakerEastWestWithLatchOpen(
       prevRobotState.modules,
-      prevRobotState.labware[labware]?.slot
+      prevRobotState.labware[labware]
     )
   ) {
     errors.push(errorCreators.heaterShakerEastWestWithLatchOpen())
   }
 
-  if (
-    getIsHeaterShakerEastWestMultiChannelPipette(
-      prevRobotState.modules,
-      prevRobotState.labware[labware]?.slot,
-      pipetteSpec
-    )
-  ) {
-    errors.push(errorCreators.heaterShakerEastWestOfMultiChannelPipette())
-  }
   if (errors.length === 0 && pipetteSpec && pipetteSpec.maxVolume < volume) {
     errors.push(
       errorCreators.pipetteVolumeExceeded({
