+++ step-generation/src/commandCreators/atomic/dispense.ts
@@ -6,7 +6,6 @@ import {
   pipetteIntoHeaterShakerWhileShaking,
   getIsHeaterShakerEastWestWithLatchOpen,
   pipetteAdjacentHeaterShakerWhileShaking,
-  getIsHeaterShakerEastWestMultiChannelPipette,
 } from '../../utils'
 import type { CreateCommand } from '@opentrons/shared-data'
 import type { DispenseParams } from '@opentrons/shared-data/protocol/types/schemaV3'
@@ -21,16 +20,6 @@ export const dispense: CommandCreator<DispenseParams> = (
   const { pipette, volume, labware, well, offsetFromBottomMm, flowRate } = args
   const actionName = 'dispense'
   const errors: CommandCreatorError[] = []
-  const pipetteSpec = invariantContext.pipetteEntities[pipette]?.spec
-
-  if (!pipetteSpec) {
-    errors.push(
-      errorCreators.pipetteDoesNotExist({
-        actionName,
-        pipette,
-      })
-    )
-  }
 
   if (
     modulePipetteCollision({
@@ -96,7 +85,7 @@ export const dispense: CommandCreator<DispenseParams> = (
   if (
     pipetteAdjacentHeaterShakerWhileShaking(
       prevRobotState.modules,
-      prevRobotState.labware[labware]?.slot
+      prevRobotState.labware[labware]
     )
   ) {
     errors.push(errorCreators.heaterShakerNorthSouthEastWestShaking())
@@ -105,20 +94,10 @@ export const dispense: CommandCreator<DispenseParams> = (
   if (
     getIsHeaterShakerEastWestWithLatchOpen(
       prevRobotState.modules,
-      prevRobotState.labware[labware]?.slot
+      prevRobotState.labware[labware]
     )
   ) {
     errors.push(errorCreators.heaterShakerEastWestWithLatchOpen())
-  }
-
-  if (
-    getIsHeaterShakerEastWestMultiChannelPipette(
-      prevRobotState.modules,
-      prevRobotState.labware[labware]?.slot,
-      pipetteSpec
-    )
-  ) {
-    errors.push(errorCreators.heaterShakerEastWestOfMultiChannelPipette())
   }
 
   if (errors.length > 0) {
