+++ step-generation/src/commandCreators/atomic/moveToWell.ts
@@ -6,7 +6,6 @@ import {
   pipetteIntoHeaterShakerWhileShaking,
   getIsHeaterShakerEastWestWithLatchOpen,
   pipetteAdjacentHeaterShakerWhileShaking,
-  getIsHeaterShakerEastWestMultiChannelPipette,
 } from '../../utils'
 import type { CreateCommand } from '@opentrons/shared-data'
 import type { MoveToWellParams as v5MoveToWellParams } from '@opentrons/shared-data/protocol/types/schemaV5'
@@ -88,7 +87,7 @@ export const moveToWell: CommandCreator<v5MoveToWellParams> = (
   if (
     pipetteAdjacentHeaterShakerWhileShaking(
       prevRobotState.modules,
-      prevRobotState.labware[labware]?.slot
+      prevRobotState.labware[labware]
     )
   ) {
     errors.push(errorCreators.heaterShakerNorthSouthEastWestShaking())
@@ -97,20 +96,10 @@ export const moveToWell: CommandCreator<v5MoveToWellParams> = (
   if (
     getIsHeaterShakerEastWestWithLatchOpen(
       prevRobotState.modules,
-      prevRobotState.labware[labware]?.slot
+      prevRobotState.labware[labware]
     )
   ) {
     errors.push(errorCreators.heaterShakerEastWestWithLatchOpen())
-  }
-
-  if (
-    getIsHeaterShakerEastWestMultiChannelPipette(
-      prevRobotState.modules,
-      prevRobotState.labware[labware]?.slot,
-      pipetteSpec
-    )
-  ) {
-    errors.push(errorCreators.heaterShakerEastWestOfMultiChannelPipette())
   }
 
   if (errors.length > 0) {
