+++ step-generation/src/types.ts
@@ -478,7 +478,6 @@ export type ErrorType =
   | 'TALL_LABWARE_EAST_WEST_OF_HEATER_SHAKER'
   | 'HEATER_SHAKER_EAST_WEST_LATCH_OPEN'
   | 'HEATER_SHAKER_NORTH_SOUTH_EAST_WEST_SHAKING'
-  | 'HEATER_SHAKER_EAST_WEST_MULTI_CHANNEL'
 
 export interface CommandCreatorError {
   message: string
