+++ step-generation/src/utils/index.ts
@@ -3,7 +3,12 @@ import { curryCommandCreator } from './curryCommandCreator'
 import { reduceCommandCreators } from './reduceCommandCreators'
 import { modulePipetteCollision } from './modulePipetteCollision'
 import { thermocyclerPipetteCollision } from './thermocyclerPipetteCollision'
+import { pipetteIntoHeaterShakerLatchOpen } from './pipetteIntoHeaterShakerLatchOpen'
+import { pipetteIntoHeaterShakerWhileShaking } from './pipetteIntoHeaterShakerWhileShaking'
 import { orderWells } from './orderWells'
+import { getIsTallLabwareEastWestOfHeaterShaker } from './getIsTallLabwareEastWestOfHeaterShaker'
+import { getIsHeaterShakerEastWestWithLatchOpen } from './getIsHeaterShakerEastWestWithLatchOpen'
+import { pipetteAdjacentHeaterShakerWhileShaking } from './pipetteAdjacentHeaterShakerWhileShaking'
 export {
   commandCreatorsTimeline,
   curryCommandCreator,
@@ -11,7 +16,11 @@ export {
   reduceCommandCreators,
   modulePipetteCollision,
   thermocyclerPipetteCollision,
+  pipetteIntoHeaterShakerLatchOpen,
+  pipetteIntoHeaterShakerWhileShaking,
+  pipetteAdjacentHeaterShakerWhileShaking,
+  getIsTallLabwareEastWestOfHeaterShaker,
+  getIsHeaterShakerEastWestWithLatchOpen,
 }
 export * from './commandCreatorArgsGetters'
 export * from './misc'
-export * from './heaterShakerCollision'
